backend/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
networkx
sqlparse
pydantic
sse-starlette
python-dotenv
pytest
httpx
